readme: Mark Container Runtimes support as ✅
CRI-O was never tested by the Confidential Containers Operator and may
be added in the future as a feature for this repo.

However, as it's not a parity situation, we can safely mark the
Container Runtimes support as ✅, and let someone expand it once more
Container Runtimes are added to the test matrix.

Signed-off-by: Fabiano Fidêncio <ffidencio@nvidia.com>